chore(components): reorganize folder and remove unused components
