--- components/LoginLogoutButton.jsx
@@ -1,45 +1,0 @@
-"use client";
-import React, { useEffect, useState } from "react";
-import { Button } from "./ui/button";
-import { useRouter } from "next/navigation";
-import { createClient } from "@/utils/supabase/client";
-import { signout } from "@/lib/auth-actions";
-
-const LoginButton = () => {
-  const [user, setUser] = useState(null);
-  const router = useRouter();
-  const supabase = createClient();
-  useEffect(() => {
-    const fetchUser = async () => {
-      const {
-        data: { user },
-      } = await supabase.auth.getUser();
-      setUser(user);
-    };
-    fetchUser();
-  }, []);
-  if (user) {
-    return (
-      <Button
-        onClick={() => {
-          signout();
-          setUser(null);
-        }}
-      >
-        Log out
-      </Button>
-    );
-  }
-  return (
-    <Button
-      variant="outline"
-      onClick={() => {
-        router.push("/login");
-      }}
-    >
-      Login
-    </Button>
-  );
-};
-
-export default LoginButton;
