+++ components/sidebar/app-sidebar.jsx
@@ -2,8 +2,8 @@
 
 import * as React from "react";
 import { Home, AlertCircle, ClipboardList, History, BookOpen, Heart, ArrowLeft } from "lucide-react";
-import { NavMain } from "@/components/nav-main";
-import { NavUser } from "@/components/nav-user";
+import { NavMain } from "@/components/sidebar/nav-main";
+import { NavUser } from "@/components/sidebar/nav-user";
 import {
   Sidebar,
   SidebarContent,
